Fixed repeated guard gasps
